backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
sqlalchemy==2.0.34
alembic==1.13.2
psycopg[binary]==3.2.13
python-dotenv==1.0.1
python-multipart==0.0.9
reportlab==4.2.0
python-docx==1.1.2
pypdf==4.3.1
PyMuPDF==1.24.9
pdfplumber==0.11.4
python-pptx==0.6.23

# OCR (scanned PDFs)
Pillow==10.4.0
pytesseract==0.3.10

# LLM (OpenAI Responses / Chat Completions)
openai>=1.30.0,<2.0.0

# Auth (optional, Mode A keeps demo headers working)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Async queue (event-driven infra)
redis==5.0.8
rq==1.16.2
